Remove fetch_data.py and test_fetch_data.py
The function get_problem_files has been relocated to the file utils/misc.py
